Cleaned up code by moving api post data set up to helper file
