+++ app/javascript/components/settings-time-profile-form/index.jsx
@@ -5,6 +5,7 @@ import PropTypes from 'prop-types';
 import createSchema from './settings-time-profile-form.schema';
 import { API } from '../../http_api';
 import miqRedirectBack from '../../helpers/miq-redirect-back';
+import { dataHelper } from './helper';
 
 const SettingsTimeProfileForm = ({
   timeProfileId, timezones, action, userid,
@@ -46,65 +47,9 @@ const SettingsTimeProfileForm = ({
     } else {
       miqSparkleOn();
 
-      let data = {};
-      let APIaction = 'edit';
-      let metrics = values.rollup_daily_metrics;
-      let profileType = values.profile_type;
-      let profile = {};
-      let { days } = values.profile;
-      const { tz } = values.profile;
-
-      if (timeProfileId === '' || action === 'timeprofile_copy') {
-        APIaction = 'create';
-      }
-
-      if (values.rollup_daily_metrics === undefined) {
-        metrics = false;
-      }
-
-      if (values.profile_type === undefined) {
-        profileType = 'user';
-      }
-
-      let hours = [];
-
-      if (values.profile.hoursAM !== undefined) {
-        values.profile.hoursAM.forEach((time) => {
-          hours.push(time);
-        });
-      }
-
-      if (values.profile.hoursPM !== undefined) {
-        values.profile.hoursPM.forEach((time) => {
-          hours.push(time);
-        });
-      }
-
-      if (values.DaysSelectAll) {
-        days = [0, 1, 2, 3, 4, 5, 6];
-      }
-
-      if (values.HoursSelectAll) {
-        hours = [0, 1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11, 12, 13, 14, 15, 16, 17, 18, 19, 20, 21, 22, 23];
-      }
-      days.sort();
-      hours.sort((int1, int2) => int1 - int2);
-
-      profile = {
-        days,
-        hours,
-        tz,
-      };
+      const data = dataHelper(values, timeProfileId, action, userid);
 
       if (timeProfileId !== '' && action !== 'timeprofile_copy') {
-        data = {
-          action: APIaction,
-          description: values.description,
-          profile_type: profileType,
-          profile_key: userid,
-          profile,
-          rollup_daily_metrics: metrics,
-        };
         const request = API.post(`/api/time_profiles/${timeProfileId}`, data);
         request
           .then(() => {
@@ -113,14 +58,6 @@ const SettingsTimeProfileForm = ({
           })
           .catch(miqSparkleOff);
       } else {
-        data = {
-          action: APIaction,
-          description: values.description,
-          profile_type: profileType,
-          profile_key: userid,
-          profile,
-          rollup_daily_metrics: metrics,
-        };
         const request = API.post(`/api/time_profiles`, data);
         request
           .then(() => {
